Added login frame for participant. Added export to csv function. Improvement changes.

+++ src/sample/BugReportController.java
@@ -5,6 +5,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.SnapshotParameters;
 import javafx.scene.control.Button;
+import javafx.scene.control.CheckBox;
 import javafx.scene.control.TextArea;
 import javafx.scene.image.WritableImage;
 import javafx.stage.Stage;
@@ -12,9 +13,9 @@ import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
 import org.apache.pdfbox.pdmodel.PDPageContentStream;
 import org.apache.pdfbox.pdmodel.common.PDRectangle;
-import org.apache.pdfbox.pdmodel.font.*;
+import org.apache.pdfbox.pdmodel.font.PDFont;
+import org.apache.pdfbox.pdmodel.font.PDType0Font;
 import org.apache.pdfbox.pdmodel.graphics.image.PDImageXObject;
-import org.apache.pdfbox.pdmodel.interactive.form.PDTextField;
 
 import javax.imageio.ImageIO;
 import java.io.File;
@@ -27,9 +28,11 @@ import java.nio.file.Paths;
 public class BugReportController {
 
     @FXML
-    Button createReport;
+    Button createReport, cancel;
     @FXML
     TextArea bugDescription;
+    @FXML
+    CheckBox needScreenshot;
 
     float POINTS_PER_INCH = 72;
     float POINTS_PER_MM = 1 / (10 * 2.54f) * POINTS_PER_INCH;
@@ -39,19 +42,10 @@ public class BugReportController {
     @FXML
     private void createReport(ActionEvent event) {
 
-        SnapshotParameters ssp = new SnapshotParameters();
-        WritableImage nodeShot = Main.getPrimaryStage().getScene().snapshot(null);
-        File file = new File("screenshot.png");
-
-        try {
-            ImageIO.write(SwingFXUtils.fromFXImage(nodeShot, null), "png", file);
-        } catch (IOException e) {
-
-        }
-
         PDDocument doc = null;
-        Path path = Paths.get("pdf_file.pdf");
-        File pdfFile = new File("pdf_file.pdf");
+        String reportName = Group.getInstance().getParticipantName() + " отчет.pdf";
+        Path path = Paths.get(reportName);
+        File pdfFile = new File(reportName);
         if (Files.exists(path)) {
             try {
                 doc = PDDocument.load(pdfFile);
@@ -63,43 +57,50 @@ public class BugReportController {
         }
 
         PDPage page = new PDPage(new PDRectangle(WIDTH_A4, HEIGHT_A4));
-        PDImageXObject pdimage;
         PDPageContentStream content;
         try {
-            pdimage = PDImageXObject.createFromFile("screenshot.png",doc);
             content = new PDPageContentStream(doc, page);
 
             content.beginText();
             //Setting the font to the Content stream
-
             PDFont formFont = PDType0Font.load(doc, new FileInputStream("times.ttf"), false);
             content.setFont(formFont, 14);
 
             //Setting the position for the line
-            content.newLineAtOffset(25 + 150, HEIGHT_A4 - 20);
-            content.showText("Пользователь: " + "      " +  "Учреждение: ");
-            content.newLineAtOffset(-150, -15);
+            content.newLineAtOffset(25 + 75, HEIGHT_A4 - 20);
+            content.showText("Пользователь: " + Group.getInstance().getParticipantName() + "      " +  "Учреждение: " + Group.getInstance().getParticipantUniversity());
+            content.newLineAtOffset(-75, -15);
             content.showText("Описание дефекта:");
             content.newLineAtOffset(0, -20);
 
-            //for testing
-            //String text = "Job Description:\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque hendrerit lectus nec ipsum gravida placerat. Fusce eu erat orci. Nunc eget augue neque. Fusce arcu risus, pulvinar eu blandit ac, congue non tellus. Sed eu neque vitae dui placerat ultricies vel vitae mi. Vivamus vulputate nullam.";
-            //String[] stringArray = text.split("\n");
             String[] stringArray = bugDescription.getText().split("\n");
             for(String paragraph : stringArray){
                 printParagraph(content, paragraph);
             }
             content.endText();
 
-            float scale = getRightScale(pdimage.getWidth(), pdimage.getHeight());
-            float rightWidth = pdimage.getWidth() / scale;
-            float rightHeight = pdimage.getHeight() / scale;
-            content.drawImage(pdimage, (WIDTH_A4 - rightWidth)/2, 20, rightWidth, rightHeight);
+            if (needScreenshot.isSelected()) {
+                SnapshotParameters ssp = new SnapshotParameters();
+                WritableImage nodeShot = Main.getPrimaryStage().getScene().snapshot(null);
+                File file = new File("screenshot.png");
+                try {
+                    ImageIO.write(SwingFXUtils.fromFXImage(nodeShot, null), "png", file);
+                } catch (IOException e) {
+                }
+
+                PDImageXObject pdimage;
+                pdimage = PDImageXObject.createFromFile("screenshot.png", doc);
+                float scale = getRightScale(pdimage.getWidth(), pdimage.getHeight());
+                float rightWidth = pdimage.getWidth() / scale;
+                float rightHeight = pdimage.getHeight() / scale;
+                content.drawImage(pdimage, (WIDTH_A4 - rightWidth) / 2, 20, rightWidth, rightHeight);
+                file.delete();
+            }
             content.close();
             doc.addPage(page);
-            doc.save("pdf_file.pdf");
+            doc.save(reportName);
             doc.close();
-            file.delete();
+
         } catch (IOException ex) {
             //Logger.getLogger(NodeToPdf.class.getName()).log(Level.SEVERE, null, ex);
         }
@@ -135,5 +136,11 @@ public class BugReportController {
         content.showText(builder.toString());
         content.newLineAtOffset(0, -15);
         builder.setLength(0);
+    }
+
+    @FXML
+    public void cancelReport(ActionEvent event) {
+        Stage stage = (Stage) cancel.getScene().getWindow();
+        stage.close();
     }
 }
